+++ 2-Tenant_Modules/12-Syllabus_Books/DDL/SyllabusBooks_ddl_v1.sql
@@ -9,7 +9,7 @@
 
 -- Authors table (Many-to-Many with Books)
 CREATE TABLE IF NOT EXISTS `bok_book_authors` (
-  `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+  `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(150) NOT NULL,
   `qualification` VARCHAR(200) DEFAULT NULL,
   `bio` TEXT DEFAULT NULL,
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS `bok_book_authors` (
 
 -- Master table for Books/Publications used across schools
 CREATE TABLE IF NOT EXISTS `bok_books` (
-  `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+  `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
   `uuid` BINARY(16) NOT NULL,  -- UUID 
   `isbn` VARCHAR(20) DEFAULT NULL,              -- International Standard Book Number
   `title` VARCHAR(100) NOT NULL,
@@ -37,9 +37,9 @@ CREATE TABLE IF NOT EXISTS `bok_books` (
   `edition` VARCHAR(50) DEFAULT NULL,           -- e.g., '5th Edition', 'Revised 2024'
   `publication_year` YEAR DEFAULT NULL,         -- e.g., 2024
   `publisher_name` VARCHAR(150) DEFAULT NULL,   -- e.g., 'NCERT', 'S.Chand', 'Pearson'
-  `language` BIGINT UNSIGNED NOT NULL,          -- FK to sys_dropdown_table e.g "English", "Hindi", "Sanskrit"
+  `language` INT UNSIGNED NOT NULL,          -- FK to sys_dropdown_table e.g "English", "Hindi", "Sanskrit"
   `total_pages` INT UNSIGNED DEFAULT NULL,
-  `cover_image_media_id` BIGINT UNSIGNED DEFAULT NULL,  -- FK to media_files.id
+  `cover_image_media_id` INT UNSIGNED DEFAULT NULL,  -- FK to media_files.id
   `tags` JSON DEFAULT NULL,                     -- Additional search tags
   `is_ncert` TINYINT(1) DEFAULT 0,              -- Flag for NCERT books
   `is_cbse_recommended` TINYINT(1) DEFAULT 0,
@@ -63,8 +63,8 @@ CREATE TABLE IF NOT EXISTS `bok_books` (
 -- ---------------------------------------------------------------------
 -- Junction: Book-Author relationship
 CREATE TABLE IF NOT EXISTS `bok_book_author_jnt` (
-  `book_id` BIGINT UNSIGNED NOT NULL,
-  `author_id` BIGINT UNSIGNED NOT NULL,
+  `book_id` INT UNSIGNED NOT NULL,
+  `author_id` INT UNSIGNED NOT NULL,
   `author_role` ENUM('PRIMARY','CO_AUTHOR','EDITOR','CONTRIBUTOR') DEFAULT 'PRIMARY',
   `ordinal` TINYINT UNSIGNED DEFAULT 1,
   PRIMARY KEY (`book_id`, `author_id`),
@@ -78,11 +78,11 @@ CREATE TABLE IF NOT EXISTS `bok_book_author_jnt` (
 -- ---------------------------------------------------------------------
 -- Link Books to Class/Subject (which books are used for which class/subject)
 CREATE TABLE IF NOT EXISTS `bok_book_class_subject_jnt` (
-  `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
-  `book_id` BIGINT UNSIGNED NOT NULL,  -- FK to slb_books.id
+  `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  `book_id` INT UNSIGNED NOT NULL,  -- FK to slb_books.id
   `class_id` INT UNSIGNED NOT NULL,    -- FK to sch_classes.id
-  `subject_id` BIGINT UNSIGNED NOT NULL, -- FK to sch_subjects.id
-  `academic_session_id` BIGINT UNSIGNED NOT NULL, -- FK to sch_org_academic_sessions_jnt.id
+  `subject_id` INT UNSIGNED NOT NULL, -- FK to sch_subjects.id
+  `academic_session_id` INT UNSIGNED NOT NULL, -- FK to sch_org_academic_sessions_jnt.id
   `is_primary` TINYINT(1) DEFAULT 1,            -- Primary textbook vs reference
   `is_mandatory` TINYINT(1) DEFAULT 1,
   `remarks` VARCHAR(255) DEFAULT NULL,
@@ -105,9 +105,9 @@ CREATE TABLE IF NOT EXISTS `bok_book_class_subject_jnt` (
 
 -- Link Book Chapters/Sections to Topics (granular mapping)
 CREATE TABLE IF NOT EXISTS `bok_book_topic_mapping` (
-  `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
-  `book_id` BIGINT UNSIGNED NOT NULL,
-  `topic_id` BIGINT UNSIGNED NOT NULL,          -- Can be topic or sub-topic at any level
+  `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  `book_id` INT UNSIGNED NOT NULL,
+  `topic_id` INT UNSIGNED NOT NULL,          -- Can be topic or sub-topic at any level
   `chapter_number` VARCHAR(20) DEFAULT NULL,    -- e.g., '1', '1.2', 'Unit I'
   `chapter_title` VARCHAR(255) DEFAULT NULL,
   `page_start` INT UNSIGNED DEFAULT NULL,
